refactor(models): move UserNotificationContent enum to enum models
