+++ crates/models/enum/src/lib.rs
@@ -679,3 +679,39 @@ pub enum NotificationPlatformLot {
     PushSafer,
     PushBullet,
 }
+
+#[derive(
+    Eq,
+    Enum,
+    Copy,
+    Clone,
+    Debug,
+    Display,
+    EnumIter,
+    PartialEq,
+    Serialize,
+    Deserialize,
+    DeriveActiveEnum,
+)]
+#[sea_orm(
+    rs_type = "String",
+    rename_all = "PascalCase",
+    db_type = "String(StringLen::None)"
+)]
+pub enum UserNotificationContent {
+    ReviewPosted,
+    MetadataPublished,
+    NewWorkoutCreated,
+    OutdatedSeenEntries,
+    MetadataStatusChanged,
+    MetadataEpisodeReleased,
+    PersonMetadataAssociated,
+    MetadataReleaseDateChanged,
+    MetadataEpisodeNameChanged,
+    MetadataEpisodeImagesChanged,
+    PersonMetadataGroupAssociated,
+    MetadataNumberOfSeasonsChanged,
+    MetadataChaptersOrEpisodesChanged,
+    NotificationFromReminderCollection,
+    IntegrationDisabledDueToTooManyErrors,
+}
